feat(campus): the public campus map is now MappedIn
`/campus/[token]/map` renders the MappedIn viewer for any campus
with a MappedIn venue (`indoorMapId`) — outdoor + indoor in one,
with `?space=` deep-linking. Campuses without a MappedIn venue fall
back to the parked Mapbox viewer.

`/campus/[token]/indoor` folds into `/map` (a redirect keeps old
links working); news + event room links now point at `/map?space=`.
